Add a feature to restrict access in Dungeons/Trials
You can now configure the behavior of Reveries when in a dungeon
or a trial.  The new options are:

1. Allow messages from everyone (normal operation)
2. Only listen to people in your group (ignore everyone else)
3. Ignore all messages

I've made it default to option 2 ("Only Group").  That way, random
guildmates or people whispering you cannot disrupt your dungeon, but
the people you're running the dungeon with can still use it.

Option 3 is essentially an auto-disable in dungeons/trials.