Add a /memento command for quick-memento activation
Activating emotes via a simple slash command is nice and convenient.

Mementos, on the other hand, require you to navigate through your
collections menu - or to quick-slot it on your ring.  Using up
quick-slot slots, and switching between trinkets and potions can
be a pain as well.  It would be nice to have slash commands.

This adds a new /memento slash command, so you can do either:

    /memento werewolf
    /memento Werewolf-Behemoth-Sigil

to activate the Werewolf Behemoth Sigil.  Tab completion works
when using LibSlashCommander as well.