Make chat channel and basic settings per-server rather than global
Kelinmiriel noted that players who play on multiple servers will have
a different set of guilds on NA vs. EU.  So, it would make more sense
for the chat channel settings to be per-server rather than global.

This patch moves the active flag, dungeon setting, and chat channel
lists to be per-server rather than global, and adds code to migrate
the existing settings over.  We leave the disabled emote and memento
lists as global, as they're likely tied to the preferences of the
account owner, rather than something server specific.