Use the full "Antiquarian's Eye" name for /scry instead of the alias
There's some kind of bug where the short aliases for mementos don't
work right, and I don't know why yet.  So, that's making the /scry
command not work sometimes.  There's no reason not to use the full
name internally, so just do that to make it a bit more reliable.