Add an on-screen cooldown bar for mementos (such as scrying)
This places a new bar over the top left area of the compass:

                 <%>-------- West Narsis Wayshrine
                 <   @         V          \%/            >

which shows the cooldown for the active memento.  In particular, this
is useful when using /scry to activate the Antiquarian's Eye.  In the
normal game, you activate it from the potion quickslot, and there's a
radial cooldown indicator for when you're able to use it again.  With
/scry however, it wasn't clear when the 30 seconds had passed and you
were able to activate it again.